requirements.txt
eventlet
pyyaml
